Fix Auto Layout issues with MeasureToolbar
The issues stemmed from the toolbar laying out its subviews before it had a valid size. The solution I used is to assign the toolbar's items in `layoutSubviews()`, which is only called after the toolbar is positioned in its super view. I also made the result view a bar item.