swift: Allow using a local install of Runtime for offline and CI
The Toolkit Swift package has a hard dependency on released versions of
the Runtime package. This amkes the Runtime very easy to setup but Swift
Packagae manager does not allow using local packages when also using a
versioned package. However, manually adding Toolkit does allow using a
local install of Runtime without the need for a repo.

This will allow developers that are offline to use Runtime with local
installations. This also allows our internal developers to use the dev
packages to test changes as well as for our CI system to run tests on
all the latest builds while still utilizing the Swift Package Manager
tooling.